refactor(reservation): ActivityStatus 위치 변경
